+++ changelog.md
@@ -1,5 +1,10 @@
 2019-07-27 - 0.9.92
 
+- toward #49 "Fully validate given config on construction"
+    - java: collect and report all "missing" errors for basic types
+      TODO handle objects
+    - TODO scala
+  
 - capture java and scala wrapper supporting methods in proper classes 
   to facilitate validation at compile time.